Add overlay toggling
