+++ packages/overlay/package.json
@@ -25,7 +25,7 @@
   "type": "module",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "types": "dist/types/index.d.ts",
   "scripts": {
     "dev": "pnpm run build --watch",
     "build": "tsx ./scripts/build.ts"
