+++ playgrounds/sandbox/src/locator.ts
@@ -1,0 +1,5 @@
+import { useLocator } from '@solid-devtools/locator'
+
+useLocator({
+  targetIDE: 'vscode',
+})
